Move master CDR function out of model.py file for sm_sir
